+++ geode-management/src/main/java/org/apache/geode/management/configuration/ClassName.java
@@ -49,7 +49,7 @@ public class ClassName implements Serializable {
   /**
    * Object to be instantiated using the empty param constructor of the className
    *
-   * @param className this class needs to a no-arg constructor.
+   * @param className this class needs a no-arg constructor.
    */
   public ClassName(String className) {
     this(className, "{}");
@@ -61,6 +61,7 @@ public class ClassName implements Serializable {
    * @param className this class needs to have a no-arg constructor
    * @param jsonInitProperties a json representation of the initialization properties
    *        that will be passed to {@link org.apache.geode.cache.Declarable#initialize}.
+   *        If the className is not Declarable, then these properties will be ignored
    * @throws IllegalArgumentException if the class name is not valid
    * @throws IllegalArgumentException if jsonInitProperties is invalid JSON
    */
@@ -80,9 +81,13 @@ public class ClassName implements Serializable {
   }
 
   /**
-   * @param className this class needs to have a no-arg constructor
+   * @param className the name of the class to be instantiated. This class needs to have
+   *        a no-arg constructor.
    * @param properties the initialization properties
    *        that will be passed to {@link org.apache.geode.cache.Declarable#initialize}.
+   *        If the className is not Declarable, then these properties will be ignored
+   *
+   * @throws IllegalArgumentException if classname contains illegal classname characters
    */
   @JsonCreator
   public ClassName(@JsonProperty("className") String className,
@@ -135,6 +140,11 @@ public class ClassName implements Serializable {
         && this.getInitProperties().equals(that.getInitProperties());
   }
 
+  /**
+   * this provides a convenient method to validate if the given name is a valid classname
+   *
+   * @return false if classname is blank or has invalid classname characters
+   */
   public static boolean isClassNameValid(String className) {
     if (StringUtils.isBlank(className)) {
       return false;
